+++ providers/secure_admin.rb
@@ -18,6 +18,7 @@ include Chef::Asadmin
 
 notifying_action :enable do
   service "glassfish-#{new_resource.domain_name}" do
+    provider Chef::Provider::Service::Upstart
     supports :restart => true
     action :nothing
   end
@@ -33,6 +34,7 @@ end
 
 notifying_action :disable do
   service "glassfish-#{new_resource.domain_name}" do
+    provider Chef::Provider::Service::Upstart
     supports :restart => true
     action :nothing
   end